Fix List.h conversion
Fix a conversion errror when try to compare int and size_t.

+++ ws3/List.h
@@ -35,7 +35,7 @@ void List<T, N>::operator+=(const T& element) {
 
 template <typename T, size_t N>
 const T& List<T, N>::operator[](int index) const {
-        if (index < 0 || index > capacity) {
+        if (index < 0 || static_cast<size_t>(index) > capacity) {
                 throw std::out_of_range("Index is out of range");
         } else {
                 return collection[index];
